refactor: split cmd/cli/main.go by command family
